Moved Choice to mol-util/param-choice.ts
