+++ src/extensions/meshes/mesh-streaming/behavior.ts
@@ -17,9 +17,9 @@ import { UUID } from '../../../mol-util';
 import { Asset } from '../../../mol-util/assets';
 import { Color } from '../../../mol-util/color';
 import { ColorNames } from '../../../mol-util/color/names';
+import { Choice } from '../../../mol-util/param-choice';
 import { ParamDefinition as PD } from '../../../mol-util/param-definition';
 
-import { Choice } from '../../volumes-and-segmentations/helpers';
 import { MetadataWrapper } from '../../volumes-and-segmentations/volseg-api/utils';
 
 import { MeshlistData } from '../mesh-extension';
