+++ src/extensions/mvs/components/selector.ts
@@ -4,7 +4,6 @@
  * @author Adam Midlik <midlik@gmail.com>
  */
 
-import { Choice } from '../../volumes-and-segmentations/helpers';
 import { SortedArray } from '../../../mol-data/int';
 import { ElementIndex, Structure, StructureElement } from '../../../mol-model/structure';
 import { StaticStructureComponentTypes, createStructureComponent } from '../../../mol-plugin-state/helpers/structure-component';
@@ -14,6 +13,7 @@ import { Expression } from '../../../mol-script/language/expression';
 import { UUID } from '../../../mol-util';
 import { arrayExtend, sortIfNeeded } from '../../../mol-util/array';
 import { mapArrayToObject, pickObjectKeys } from '../../../mol-util/object';
+import { Choice } from '../../../mol-util/param-choice';
 import { ParamDefinition as PD } from '../../../mol-util/param-definition';
 import { capitalize } from '../../../mol-util/string';
 import { MVSAnnotationStructureComponentParams, createMVSAnnotationStructureComponent } from './annotation-structure-component';
